+++ UnityUtils/Assets/Testing Scripts/TestingScript.cs
@@ -6,7 +6,7 @@ public class TestingScript : MonoBehaviour {
   public float testingValue;
   public Vector3 testingVector3;
 
-  public SFX[] sfx;
+  public SFX sfx;
   // Starter is called once before the first execution of Update after the MonoBehaviour is created
   void Start() {
     ObjectAnimations.AnimateTransformRotation(transform, new Vector3(transform.localRotation.x, transform.localRotation.y, transform.localRotation.z), new Vector3(4, 50, 90), 2);