+++ UnityUtils/Packages/com.mydimons.unityutils/Runtime/Script Utils/UI/UIButtonSfx.cs
@@ -11,46 +11,21 @@ namespace UnityUtils.ScriptUtils.UI {
     /// Sound to play when the button is hovered over
     /// </summary>
     [Header("Audio Clips & Volumes")]
-    public AudioClip hoverEnterSfx;
-    /// <summary>
-    /// Volume to play <see cref="hoverEnterSfx"/> at.
-    /// </summary>
-    public float hoverEnterVolume = AudioManager.MAX_AUDIO_VOLUME;
+    public SFX hoverEnterSfx;
 
     [Space(5)]
 
     /// <summary>
     /// Sound to play when the button is no longer being hovered.
     /// </summary>
-    public AudioClip hoverExitSfx;
-    /// <summary>
-    /// Volume to play <see cref="hoverExitSfx"/> at.
-    /// </summary>
-    public float hoverExitVolume = AudioManager.MAX_AUDIO_VOLUME;
+    public SFX hoverExitSfx;
 
     [Space(5)]
 
     /// <summary>
     /// Sound to play when the button is clicked
     /// </summary>
-    public AudioClip clickSfx;
-
-    /// <summary>
-    /// Volume to play <see cref="clickSfx"/> at.
-    /// </summary>
-    public float clickVolume = AudioManager.MAX_AUDIO_VOLUME;
-
-
-    /// <summary>
-    /// Random variance to play all audio clips at.
-    /// </summary>
-    [Header("Adjustable Values")]
-    public float pitchVariance = AudioManager.DEFAULT_PITCH_VARIANCE;
-
-    /// <summary>
-    /// Type of audio to use to get proper audio levels.
-    /// </summary>
-    public AudioManager.VolumeType volumeType = AudioManager.VolumeType.Sfx;
+    public SFX clickSfx;
 
     private bool sceneLoadTriggered;
 
@@ -62,7 +37,7 @@ namespace UnityUtils.ScriptUtils.UI {
 
     public void OnPointerEnter(PointerEventData eventData) {
       if (hoverEnterSfx != null)
-        SfxManager.PlaySfxAudioClip(hoverEnterSfx, hoverEnterVolume, pitchVariance, volumeType);
+        SfxManager.PlaySFX(hoverEnterSfx);
       else if (logEmptySfx)
         Debug.LogWarning("No hover enter SFX on button!");
     }
@@ -72,7 +47,7 @@ namespace UnityUtils.ScriptUtils.UI {
         return;
 
       if (hoverExitSfx != null)
-        SfxManager.PlaySfxAudioClip(hoverExitSfx, hoverExitVolume, pitchVariance, volumeType);
+        SfxManager.PlaySFX(hoverExitSfx);
       else if (logEmptySfx)
         Debug.LogWarning("No hover exit SFX on button!");
     }
@@ -84,7 +59,7 @@ namespace UnityUtils.ScriptUtils.UI {
       }
 
       if (clickSfx != null)
-        SfxManager.PlaySfxAudioClip(clickSfx, clickVolume, pitchVariance, volumeType);
+        SfxManager.PlaySFX(clickSfx);
       else if (logEmptySfx)
         Debug.LogWarning("No click SFX on button!");
     }